+++ src-bex/content/floatplane.ts
@@ -44,6 +44,39 @@ export default function(bridge: BexBridge) {
         )
       } else if (isChannelPage(url)) {
         // Channel Page Handler
+        const progressdata = (await bridge.send('video.getprogress'))?.data || {}
+        videoFeedObserver = watchForElement<HTMLDivElement>(
+          FEED_ELEMENT,
+          async (videoFeedEl) => {
+            if (!settings.progressTracking || !videoFeedEl) return
+
+            const items = videoFeedEl.querySelectorAll('div.ReactElementGridItem')
+            items.forEach((itemEl) => {
+              const linkEl = itemEl.querySelector<HTMLAnchorElement>('div.PostTileWrapper > a')
+              if (!linkEl) return
+
+              const videoId = getVideoId(linkEl.href)
+              const videoProgress = progressdata[videoId]
+
+              // Skip item if we've already injected the watch indicator
+              if (!videoProgress || itemEl.classList.contains('watched')) {
+                return
+              }
+
+              const thumbEl = itemEl.querySelector<HTMLDivElement>('div.PostTileThumbnail')
+              const durationEl = itemEl.querySelector<HTMLDivElement>('div.duration > .text')
+              if (!thumbEl || !durationEl) return
+
+              const duration = parseDuration(durationEl.textContent || '')
+              thumbEl.appendChild(createProgressBar(videoProgress.progress / duration))
+              itemEl.classList.add('watched')
+
+              if (videoProgress.progress / duration >= 0.95) {
+                itemEl.classList.add('watched-fully')
+              }
+            })
+          }
+        )
       }
     })
   }
@@ -60,4 +93,40 @@ function getVideoId(url: string) {
 
 function isChannelPage(url: string) {
   return /^https:\/\/(www|beta)\.floatplane.com\/channel/.test(url)
+}
+
+function parseDuration(durationStr: string): number {
+  if (!durationStr) return 0
+
+  const parts = durationStr.split(':')
+  if (parts.length === 3) {
+    return (parseInt(parts[0]) * 60 * 60) +
+      (parseInt(parts[1]) * 60) +
+      parseInt(parts[2])
+  } else if (parts.length === 2) {
+    return (parseInt(parts[0]) * 60) +
+      parseInt(parts[1])
+  } else {
+    return parseInt(durationStr)
+  }
+}
+
+function createProgressBar(progress: number): HTMLDivElement {
+  const backgroundEl = document.createElement('div')
+  backgroundEl.style.position = 'absolute'
+  backgroundEl.style.height = '8px'
+  backgroundEl.style.background = 'rgba(0, 0, 0, 0.35)'
+  backgroundEl.style.bottom = '0'
+  backgroundEl.style.left = '0'
+  backgroundEl.style.right = '0'
+  backgroundEl.style.zIndex = '1'
+  backgroundEl.style.borderTop = '1px solid black'
+
+  const progressEl = document.createElement('div')
+  progressEl.style.width = `${Math.round(progress * 100)}%`
+  progressEl.style.height = '100%'
+  progressEl.style.background = '#0085ff'
+
+  backgroundEl.appendChild(progressEl)
+  return backgroundEl
 }