+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add tooltip to search results' title
 - Automatically resume video from where the user left off
+- Display watch progress on Channel Page
 
 ### Fixed
 - Handle posts without thumbnail properly